requirements.txt
fastapi
uvicorn[standard]
pandas
jieba
requests
python-dotenv
pymongo
azure-ai-textanalytics